+++ CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
 * Add go module
 * Remove iojs support
 * Restore curl util to fix -1 content length
+* Fix npm mirror url, add tuna, opentuna and ustc
 
 - 2016-04-27, Version 0.2.0:
 * Add new feature `gnvm session` Use any Node.js version of the local already exists version by current session.